fix: add validatte body formats

+++ public/src/controllers/roomController.ts
@@ -13,21 +13,51 @@ export class RoomController {
     async createRoom(req: Request, res: Response): Promise<void> {
         try {
             const { building_id, manager_id, name, schedule, workspace_config, equipments } = req.body;
-            if (!building_id || !manager_id || !name || !schedule || !workspace_config || !equipments) {
-                res.status(400).json({ message: "Dados inválidos. Todos os campos são obrigatórios." });
+
+            if (!building_id) {
+                res.status(400).json({ message: "O campo id do prédio é obrigatório." });
                 return;
             }
-    
-            const room = await this.roomService.createRoom(building_id, manager_id, name, schedule, workspace_config, equipments );
+            if (isNaN(Number(building_id))) {
+                res.status(400).json({ error: "O id do prédio deve ser um número" });
+                return 
+            }
+            if (!manager_id) {
+                res.status(400).json({ message: "O campo id do gerente é obrigatório." });
+                return;
+            }
 
-            if(!room) {
+            const cpfRegex = /^\d{3}\.\d{3}\.\d{3}\-\d{2}$/;
+            if (!cpfRegex.test(manager_id)) {
+                res.status(400).json({ message: "O campo id do gerente deve estar no formato XXX.XXX.XXX-XX." });
+                return;
+            }
+            if (!name) {
+                res.status(400).json({ message: "O campo nome é obrigatório." });
+                return;
+            }
+            if (!schedule) {
+                res.status(400).json({ message: "O campo horários é obrigatório." });
+                return;
+            }
+            if (!workspace_config) {
+                res.status(400).json({ message: "O campo configuração de espaço é obrigatório." });
+                return;
+            }
+            if (!equipments) {
+                res.status(400).json({ message: "O campo equipamentos é obrigatório." });
+                return;
+            }
+
+            const room = await this.roomService.createRoom(building_id, manager_id, name, schedule, workspace_config, equipments);
+
+            if (!room) {
                 res.status(409).json({ message: "Uma sala com este nome já existe." });
                 return;
             }
 
             res.status(201).json(room);
         } catch (error: any) {
-
             if (error.message === "Id do prédio não encontrado") {
                 res.status(404).json({ message: error.message });
                 return;
@@ -38,7 +68,7 @@ export class RoomController {
             }
 
             res.status(500).json({ message: "Erro ao criar a sala", error: error.message });
-    }
+        }
     };
 
     //app.get("/building/:building_id", (req, res) =>
@@ -46,21 +76,24 @@ export class RoomController {
     async getRoomsByBuilding(req: Request, res: Response): Promise<void> {
         try {
             const { building_id } = req.params;
-            const rooms = await this.roomService.getRoomsByBuilding(Number(building_id));
 
+            if (!building_id) {
+                res.status(400).json({ message: "O campo id do prédio é obrigatório." });
+                return;
+            }
             if (isNaN(Number(building_id))) {
-                res.status(400).json({ error: "O ID do prédio deve ser um número." });
+                res.status(400).json({ error: "O id do prédio deve ser um número." });
                 return 
             }
 
-            if (!rooms) {
+            const rooms = await this.roomService.getRoomsByBuilding(Number(building_id));
 
+            if (!rooms) {
                 res.status(404).json({ message: "Prédio não encontrado" });
                 return;
             }
             if (rooms.length === 0) {
                 res.status(204).json({ message: "Não há salas resgistradas" }); 
-
                 return;
             }
 
@@ -75,6 +108,18 @@ export class RoomController {
     async getRoomsByManager(req: Request, res: Response): Promise<void> {
         try {
             const { manager_id } = req.params;
+
+            if (!manager_id) {
+                res.status(400).json({ message: "O campo id do gerente é obrigatório." });
+                return;
+            }  
+
+            const cpfRegex = /^\d{3}\.\d{3}\.\d{3}\-\d{2}$/;
+            if (!cpfRegex.test(manager_id)) {
+                res.status(400).json({ message: "O campo id do gerente deve estar no formato XXX.XXX.XXX-XX." });
+                return;
+            }
+
             const rooms = await this.roomService.getRoomsByManager(manager_id);
 
             if (!rooms) {
@@ -97,15 +142,21 @@ export class RoomController {
     async getOneRoom(req: Request, res: Response): Promise<void> {
         try {
             const { id } = req.params;
-            const rooms = await this.roomService.getOneRoom(Number(id));
 
-            if (isNaN(Number(id))) {
-                res.status(400).json({ error: "O ID da sala deve ser um número" });
-                return 
+            if (!id) {
+                res.status(400).json({ message: "O campo id da sala é obrigatório." });
+                return;
             }
 
+            if (isNaN(Number(id))) {
+                res.status(400).json({ error: "O id da sala deve ser um número" });
+                return 
+            }
+            
+            const rooms = await this.roomService.getOneRoom(Number(id));
+
             if (!rooms) {
-                res.status(404).json({ message: `Sala com ID ${id} não encontrada`});
+                res.status(404).json({ message: `Sala com id ${id} não encontrada`});
                 return;
             }
 
@@ -120,19 +171,61 @@ export class RoomController {
     async editOneRoom(req: Request, res: Response): Promise<void> {
         try {
             const { id } = req.params;
-            const { building_id, manager_id, name, schedule, workspace_config, equipments } = req.body;
 
-            if (!building_id || !manager_id || !name || !schedule || !workspace_config || !equipments ) {
-                res.status(400).json({ message: "Dados inválidos. Todos os campos são obrigatórios" });
-                return
-            }
-    
-            const room = await this.roomService.editOneRoom(Number(id), building_id, manager_id, name, schedule, workspace_config, equipments);
-
-            if(!room) {
-                res.status(404).json({ message: `Sala com ID ${id} não encontrada` });
+            if (!id) {
+                res.status(400).json({ message: "O campo id da sala é obrigatório." });
                 return;
             }
+
+            if (isNaN(Number(id))) {
+                res.status(400).json({ error: "O id da sala deve ser um número" });
+                return 
+            }
+
+            const { building_id, manager_id, name, schedule, workspace_config, equipments } = req.body;
+
+            if (!building_id) {
+                res.status(400).json({ message: "O campo id do prédio é obrigatório." });
+                return;
+            }
+            if (isNaN(Number(building_id))) {
+                res.status(400).json({ error: "O id do prédio deve ser um número" });
+                return 
+            }
+            if (!manager_id) {
+                res.status(400).json({ message: "O campo id do gerente é obrigatório." });
+                return;
+            }
+
+            const cpfRegex = /^\d{3}\.\d{3}\.\d{3}\-\d{2}$/;
+            if (!cpfRegex.test(manager_id)) {
+                res.status(400).json({ message: "O campo id do gerente deve estar no formato XXX.XXX.XXX-XX." });
+                return;
+            }
+            if (!name) {
+                res.status(400).json({ message: "O campo nome é obrigatório." });
+                return;
+            }
+            if (!schedule) {
+                res.status(400).json({ message: "O campo horários é obrigatório." });
+                return;
+            }
+            if (!workspace_config) {
+                res.status(400).json({ message: "O campo configuração de espaço é obrigatório." });
+                return;
+            }
+            if (!equipments) {
+                res.status(400).json({ message: "O campo equipamentos é obrigatório." });
+                return;
+            }
+
+            const room = await this.roomService.editOneRoom(Number(id), building_id, manager_id, name, schedule, workspace_config, equipments);
+
+            if (!room) {
+                res.status(404).json({ message: `Sala com id ${id} não encontrada` });
+                return;
+            }
+            
 
             res.status(201).json(room);
         } catch (error: any) {
@@ -145,21 +238,32 @@ export class RoomController {
                 return;
             }
             res.status(500).json({ message: "Erro ao editar a sala", error: error.message });
-    }
+        }
     };
     
     //app.delete("/:id", (req, res) =>
     
     async deleteOneRoom(req: Request, res: Response): Promise<void> {
         try {
-            const { id } = req.params; 
+            const { id } = req.params;
+
+            if (!id) {
+                res.status(400).json({ message: "O campo id da sala é obrigatório." });
+                return;
+            }
+
+            if (isNaN(Number(id))) {
+                res.status(400).json({ error: "O id da sala deve ser um número" });
+                return 
+            }
+
             const deleted = await this.roomService.deleteOneRoom(Number(id));
     
             if (deleted) {
-                res.status(204).json({ message: `Sala com ID ${id} excluída com sucesso.` });
+                res.status(204).json({ message: `Sala com id ${id} excluída com sucesso.` });
                 return;
             } else {
-                res.status(404).json({ message: `Sala com ID ${id} não encontrada.` });
+                res.status(404).json({ message: `Sala com id ${id} não encontrada.` });
                 return;
             }
